Release 0.2.0 (#150)
Release Notes

* Bumped haze, agp, and guava dependencies
* Implemented ordering functionality for friends list
* Added scroll to top feature in friends and conversations screens
* Improved messages handling
* Fixed coloring issues
* Cache improvements
* Implemented logout functionality
* Implemented new authorization flow (no auto-token re-request)
* Added support for sticker pack preview attachments
* Bump LongPoll to version 19
* Markdown support for messages bubbles
* Adjust app name font size based on screen width

---------

+++ core/data/src/main/kotlin/dev/meloda/fast/data/VkGroupsMap.kt
@@ -10,7 +10,7 @@ class VkGroupsMap(
     private val groups: List<VkGroupDomain>
 ) {
 
-    private val map: HashMap<Int, VkGroupDomain> by lazy {
+    private val map: HashMap<Long, VkGroupDomain> by lazy {
         HashMap(groups.associateBy(VkGroupDomain::id))
     }
 
@@ -36,7 +36,7 @@ class VkGroupsMap(
         if (message.fromId >= 0) null
         else map[abs(message.fromId)]
 
-    fun group(groupId: Int): VkGroupDomain? = map[abs(groupId)]
+    fun group(groupId: Long): VkGroupDomain? = map[abs(groupId)]
 
     companion object {
