+++ core/model/src/main/kotlin/dev/meloda/fast/model/FriendsInfo.kt
@@ -4,5 +4,5 @@ import dev.meloda.fast.model.api.domain.VkUser
 
 data class FriendsInfo(
     val friends: List<VkUser>,
-    val onlineFriendsIds: List<Int>
+    val onlineFriendsIds: List<Long>
 )
