+++ core/model/src/main/kotlin/dev/meloda/fast/model/api/data/VkGraffitiData.kt
@@ -1,13 +1,13 @@
 package dev.meloda.fast.model.api.data
 
-import dev.meloda.fast.model.api.domain.VkGraffitiDomain
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import dev.meloda.fast.model.api.domain.VkGraffitiDomain
 
 @JsonClass(generateAdapter = true)
 data class VkGraffitiData(
-    @Json(name = "id") val id: Int,
-    @Json(name = "owner_id") val ownerId: Int,
+    @Json(name = "id") val id: Long,
+    @Json(name = "owner_id") val ownerId: Long,
     @Json(name = "url") val url: String,
     @Json(name = "width") val width: Int,
     @Json(name = "height") val height: Int,
