+++ core/model/src/main/kotlin/dev/meloda/fast/model/api/domain/VkGraffitiDomain.kt
@@ -3,8 +3,8 @@ package dev.meloda.fast.model.api.domain
 import dev.meloda.fast.model.api.data.AttachmentType
 
 data class VkGraffitiDomain(
-    val id: Int,
-    val ownerId: Int,
+    val id: Long,
+    val ownerId: Long,
     val url: String,
     val width: Int,
     val height: Int,
